feat(connection list): list connection databases and store current connection/database
Currently, the user wasn't able to connect to a database and the current connection and database
weren't stored in the application state.

Now the user can connect to the saved connections in the
sidebar and list the databases inside the connection. The user can also connect to a single
database.

fix #34, fix #35